Link styled text, move and rename Incident components
